% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_msgf_output_to_msnid.R
\name{convert_msgf_output_to_msnid}
\alias{convert_msgf_output_to_msnid}
\title{Converting MS-GF+ output to MSnID}
\usage{
convert_msgf_output_to_msnid(x)
}
\arguments{
\item{results}{(data.frame) concatenated output of MS-GF+}
}
\value{
(MSnID) MSnID object
}
\description{
Converting MS-GF+ output to MSnID
}
