% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_links_from_single_job.R
\name{get_links_from_single_job}
\alias{get_links_from_single_job}
\title{Finding Links from PNNL's DMS}
\usage{
get_links_from_single_job(scansStats)
}
\arguments{
\item{pathToFile}{}
}
\value{
(data.table) with reporter ion intensities and other metrics
}
\description{
Finding Links from PNNL's DMS
}
