% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_id_to_quant_scan_link_table.R
\name{make_id_to_quant_scan_link_table}
\alias{make_id_to_quant_scan_link_table}
\alias{make_id_to_quant_scan_link_table_for_dataset}
\title{Finding Links from PNNL's DMS}
\usage{
make_id_to_quant_scan_link_table(
  scanStats,
  idScanPattern = "ms2",
  quantScanPattern = "ms3",
  threshold = NULL
)

make_id_to_quant_scan_link_table_for_dataset(
  scanStats,
  quantScanPattern,
  idScanPattern,
  threshold = NULL
)
}
\arguments{
\item{scanStats}{placeholder}

\item{idScanPattern}{placeholder}

\item{quantScanPattern}{placeholder}

\item{threshold}{placeholder}
}
\value{
(data.table) with reporter ion intensities and other metrics
}
\description{
Link ID scans to quantification scans.
* `make_id_to_quant_scan_link_table()`: Link ID scans to quant scans for 
    multiple datasets.
* `make_id_to_quant_scan_link_table_for_dataset()`: Link ID scans to quant
    scans for a single dataset.
}
