% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PNNL_DMS_utils.R,
%   R/read_AScore_results_from_DMS.R
\docType{data}
\name{pnnl_dms_utils}
\alias{pnnl_dms_utils}
\alias{tool2suffix}
\alias{is_PNNL_DMS_connection_successful}
\alias{get_dms_job_records}
\alias{get_tool_output_files_for_job_number}
\alias{get_output_folder_for_job_and_tool}
\alias{get_AScore_results}
\alias{get_job_records_by_dataset_package}
\alias{get_datasets_by_data_package}
\alias{download_datasets_by_data_package}
\alias{get_results_for_multiple_jobs}
\alias{get_results_for_multiple_jobs.dt}
\alias{get_results_for_single_job}
\alias{get_url_from_dir_and_file}
\alias{get_results_for_single_job.dt}
\alias{path_to_FASTA_used_by_DMS}
\alias{read_AScore_results_from_DMS}
\title{Utilities for interacting with PNNL DMS}
\format{
An object of class \code{list} of length 7.
}
\usage{
tool2suffix

is_PNNL_DMS_connection_successful()

get_dms_job_records(
  jobs = NULL,
  datasetPttrn = "",
  experimentPttrn = "",
  toolPttrn = "",
  parPttrn = "",
  settingsPttrn = "",
  fastaPttrn = "",
  proteinOptionsPttrn = "",
  instrumentPttrn = ""
)

get_tool_output_files_for_job_number(
  jobNumber,
  toolName = NULL,
  fileNamePttrn,
  mostRecent = TRUE
)

get_output_folder_for_job_and_tool(jobNumber, toolName, mostRecent = TRUE)

get_AScore_results(data_package_num)

get_job_records_by_dataset_package(data_package_num)

get_datasets_by_data_package(data_package_num)

download_datasets_by_data_package(
  data_package_num,
  copy_to = ".",
  fileNamePttrn = ".raw"
)

get_results_for_multiple_jobs(jobRecords)

get_results_for_multiple_jobs.dt(jobRecords, expected_multiple_files = FALSE)

get_results_for_single_job(pathToFile, fileNamePttrn)

get_url_from_dir_and_file(dir, file_name_segment)

get_results_for_single_job.dt(
  pathToFile,
  fileNamePttrn,
  expected_multiple_files = FALSE
)

path_to_FASTA_used_by_DMS(data_package_num, organism_db = NULL)

read_AScore_results_from_DMS(data_package_num)
}
\arguments{
\item{jobs}{(integer) DMS job ID}

\item{datasetPttrn}{placeholder}

\item{experimentPttrn}{placeholder}

\item{toolPttrn}{placeholder}

\item{parPttrn}{placeholder}

\item{settingsPttrn}{placeholder}

\item{fastaPttrn}{placeholder}

\item{proteinOptionsPttrn}{placeholder}

\item{instrumentPttrn}{placeholder}

\item{jobNumber}{placeholder}

\item{toolName}{placeholder}

\item{fileNamePttrn}{placeholder}

\item{mostRecent}{(logical) only most recent or all output files}

\item{data_package_num}{(Numeric or Character vector) containing Data Package ID(s) located in DMS}

\item{jobRecords}{placeholder}

\item{expected_multiple_files}{do we expect multiple files for a single job or not. Default is FALSE.}

\item{pathToFile}{placeholder}

\item{dir}{placeholder}

\item{file_name_segment}{placeholder}

\item{organism_db}{(character) FASTA file. This is the same as the
\code{Organism DB} column. No need to specify this if there is only
one FASTA file associated with the jobs.}
}
\value{
ascore (data.frame) AScore results
}
\description{
Fetches results of different data analysis tools.
Not useful outside of PNNL unless connected through VPN.
Works on windows out of the box. To make it work on Mac/Linux FreeTDS needs
to be installed. Mac/Linux functionality is not well tested.

\itemize{
\item \code{get_dms_job_records()}: returns job records given variety keyword patterns
\item \code{get_tool_output_files_for_job_number()}: returns the output of the tool given tool name and file pattern
\item \code{get_output_folder_for_job_and_tool()}: returns the path given job id and tool name
\item \code{get_AScore_results()}: returns Ascore results given data package number
\item \code{get_job_records_by_dataset_package()}: returns job records given data package number
\item \code{get_results_for_multiple_jobs()}: returns concatenated results given job numbers
\item \code{get_results_for_multiple_jobs.dt()}: returns results as concatenated data.table given job numbers
\item \code{get_results_for_single_job()}: returns results given job number
\item \code{get_results_for_single_job.dt()}: returns results as data.table given job number
\item \code{get_datasets_by_data_package()}: returns table with datasets info
\item \code{download_datasets_by_data_package()}: downloads datasets into provided location
}

This function read AScore results from a local folder.
}
\examples{
get_output_folder_for_job_and_tool(863951, "DTA_Refinery")
data_package_num <- 3625
# msnid <- read_msgf_data_from_DMS(data_package_num) # Not run
ascore <- read_AScore_results_from_DMS(data_package_num)
# Use PlexedPiper to make use of the Ascore results
# msnid <- PlexedPiper::best_PTM_location_by_ascore(msnid, ascore)
}
\keyword{datasets}
