% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_AScore_results_from_DMS.R
\name{read_AScore_results_from_DMS}
\alias{read_AScore_results_from_DMS}
\title{Read AScore results from folder}
\usage{
read_AScore_results_from_DMS(data_package_num)
}
\arguments{
\item{data_package_num}{(Numeric or Character vector) containing Data Package
ID(s) located in DMS}
}
\value{
ascore (data.frame) AScore results
}
\description{
This function read AScore results from a local folder.
}
\examples{
\dontrun{
if (is_PNNL_DMS_connection_successful()) {
  data_package_num <- 3625
  # msnid <- read_msgf_data_from_DMS(data_package_num) # Not run
  ascore <- read_AScore_results_from_DMS(data_package_num)
  # Use PlexedPiper to make use of the Ascore results
  # msnid <- PlexedPiper::best_PTM_location_by_ascore(msnid, ascore)
}
}

}
