% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_genetic_tool_results_from_DMS.R
\name{read_generic_job_from_DMS}
\alias{read_generic_job_from_DMS}
\title{Reading generic tool output from PNNL's DMS}
\usage{
read_generic_job_from_DMS(data_package_num, tool)
}
\arguments{
\item{data_package_num}{numeric or character; the Data Package ID in the DMS.}

\item{tool}{character; MS-GF+ parameter file. No need to specify this
if there is only one parameter file associated with the jobs.}
}
\value{
(data.table) results in the form of one data.table
}
\description{
Reading generic tool output from PNNL's DMS
}
\examples{
if (is_PNNL_DMS_connection_successful()) {
  x <- read_generic_job_from_DMS(6148, "Decon2LS_V2")
  (x)
}

}
