% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_msgf_data_from_DMS.R
\name{read_msgf_data_from_DMS}
\alias{read_msgf_data_from_DMS}
\title{Reading MSGF output from PNNL's DMS}
\usage{
read_msgf_data_from_DMS(data_package_num)
}
\arguments{
\item{data_package_num}{(Numeric or Character vector) containing Data Package ID(s) located in DMS}
}
\value{
(data.frame) concatenated job results
}
\description{
Reading MSGF output from PNNL's DMS
}
\examples{
results <- read_msgf_data_from_DMS(3442)
head(results)
}
