% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_MSstats_from_MSFragger_job.R
\name{read_MSstats_from_MSFragger_job}
\alias{read_MSstats_from_MSFragger_job}
\title{Reading MSFragger-generated MSstats from PNNL's DMS as MSnSet object}
\usage{
read_MSstats_from_MSFragger_job(
  data_package_num,
  param_file = NULL,
  settings_file = NULL,
  organism_db = NULL
)
}
\arguments{
\item{data_package_num}{numeric or character; the Data Package ID in the DMS.}

\item{param_file}{character; MSFragger parameter file. No need to specify
this if there is only one parameter file associated with the jobs.}

\item{settings_file}{character; MSFragger settings file. No need to specify
this if there is only one settings file associated with the jobs.}

\item{organism_db}{character; FASTA file name. This is the same as the
\code{Organism DB} column. No need to specify this if there is only one
FASTA file associated with the jobs.}
}
\value{
(MSnSet) MSnSet object
}
\description{
Function has only been tested with label-free intensity-based
  quantification data. It has not been tested with TMT data. MSstats.csv is
  an optional output file which needs to be specified in DMS settings file.
}
\examples{
if (is_PNNL_DMS_connection_successful()) {
  msnset <- read_MSstats_from_MSFragger_job(4139)
  show(msnset)
}

}
