% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_study_design_from_DMS.R
\name{read_study_design}
\alias{read_study_design}
\alias{path_to_study_design_from_DMS}
\title{Utilities for reading study design files.}
\usage{
path_to_study_design_from_DMS(data_package_num)
}
\arguments{
\item{data_package_num}{(integer) data package number for DMS}
}
\description{
Fetches study design results from DMS.
Checks that study design files are internally consistent.
DMS functionality not useful outside of PNNL unless connected through VPN.

* `read_study_design_from_DMS()`: finds data package folder in DMS and calls read_study_design there
}
\examples{
study_design <- read_study_design_from_DMS(3606)

fractions  <- study_design$fractions
samples    <- study_design$samples
references <- study_design$references
}
