% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_masic_data_from_DMS.R
\name{read_masic_data_from_DMS}
\alias{read_masic_data_from_DMS}
\title{Reading MASIC results from PNNL's DMS}
\usage{
read_masic_data_from_DMS(
  data_package_num,
  interference_score = FALSE,
  id_quant_table = FALSE
)
}
\arguments{
\item{data_package_num}{(numeric or character vector) containing Data Package 
ID(s) located in DMS.}

\item{interference_score}{(logical) read interference score. Default is 
\code{FALSE}.}

\item{id_quant_table}{If the MS/MS identification scans and reporter ion 
intensity scans are different for a particular dataset, set this to 
\code{TRUE}. Default is \code{FALSE}.}
}
\value{
(data.frame) with reporter ion intensities and other metrics
}
\description{
Reading MASIC results from PNNL's DMS
}
