% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_masic_data_from_DMS.R
\name{read_masic_data_from_DMS}
\alias{read_masic_data_from_DMS}
\title{Reading MASIC results from PNNL's DMS}
\usage{
read_masic_data_from_DMS(
  data_package_num,
  interference_score = FALSE,
  idScanPattern = "ms2",
  quantScanPattern = "ms2"
)
}
\arguments{
\item{data_package_num}{(numeric or character vector) containing Data Package
ID(s) located in DMS.}

\item{interference_score}{(logical) read interference score. Default is
\code{FALSE}.}

\item{quantScanPattern}{Ion fragmentation method used to produce the scan used for 
peptide quantification}

\item{idscanpattern}{Ion fragmentation method used to produce the scan used for 
peptide identification}
}
\value{
(data.frame) with reporter ion intensities and other metrics
}
\description{
Reading MASIC results from PNNL's DMS
}
\examples{
if (is_PNNL_DMS_connection_successful()) {
test <- read_masic_data_from_DMS(4970)
}
}
