% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_masic_data_from_DMS.R
\name{read_masic_data_from_DMS}
\alias{read_masic_data_from_DMS}
\title{Reading MASIC results from PNNL's DMS}
\usage{
read_masic_data_from_DMS(
  dataPkg,
  interferenceScore = FALSE,
  idQuantTable = FALSE
)
}
\arguments{
\item{dataPkg}{(Numeric or Character vector) containing Data Package ID(s) located in DMS}

\item{interferenceScore}{(logical) read interference score. Default is FALSE.}
}
\value{
(data.frame) with reporter ion intensities and other metrics
}
\description{
Reading MASIC results from PNNL's DMS
}
