% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_msfragger_data_from_DMS.R
\name{read_msfragger_data_from_DMS}
\alias{read_msfragger_data_from_DMS}
\title{Reading MSFragger output from PNNL's DMS}
\usage{
read_msfragger_data_from_DMS(
  data_package_num,
  param_file = NULL,
  settings_file = NULL,
  organism_db = NULL
)
}
\arguments{
\item{data_package_num}{numeric or character; the Data Package ID in the DMS.}

\item{param_file}{character; MSFragger parameter file. No need to specify this
if there is only one parameter file associated with the jobs.}

\item{settings_file}{character; MSFragger settings file. No need to specify this
if there is only one settings file associated with the jobs.}

\item{organism_db}{character; FASTA file name. This is the same as the
\code{Organism DB} column. No need to specify this if there is only one
FASTA file associated with the jobs.}
}
\value{
(MSnID) MSnID object
}
\description{
Reading MSFragger output from PNNL's DMS
}
\examples{
msnid <- read_msfragger_data_from_DMS(4139)
show(msnid)

}
