% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_msgf_data_from_DMS.R
\name{read_msgf_data_from_DMS}
\alias{read_msgf_data_from_DMS}
\title{Reading MS-GF+ output from PNNL's DMS}
\usage{
read_msgf_data_from_DMS(
  data_package_num,
  param_file = NULL,
  organism_db = NULL
)
}
\arguments{
\item{data_package_num}{(Numeric or Character vector) containing Data 
Package ID(s) located in DMS.}

\item{param_file}{(character) MS-GF+ parameter file. No need to specify this 
if there is only one parameter file associated with the jobs.}

\item{organism_db}{(character) FASTA file. This is the same as the 
\code{Organism DB} column. No need to specify this if there is only 
one FASTA file associated with the jobs.}
}
\value{
(MSnID) MSnID object
}
\description{
Reading MS-GF+ output from PNNL's DMS
}
\examples{
msnid <- read_msgf_data_from_DMS(3606)
show(msnid)
}
